requirements.txt
absl-py==0.15.0
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
argh==0.26.2
arrow==0.13.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
black==19.10b0
bleach==4.0.0
boto3==1.19.9
botocore==1.22.9
brotlipy==0.7.0
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
configparser==5.0.2
cookiecutter==1.7.2
cryptography==3.4.8
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
Faker==9.5.2
filelock==3.3.1
firebase-admin==5.0.3
flake8==3.9.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flask-restplus==0.13.0
flask-restx==0.5.1
flatbuffers==1.12
gast==0.4.0
gcloud==0.18.3
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
h5py==3.1.0
httplib2==0.20.1
huggingface-hub==0.0.19
idna==3.2
ijson==3.1.4
imagesize==1.2.0
importlib-metadata==4.8.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jwcrypto==1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==23.1.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
numpy==1.21.4
numpydoc==1.1.0
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.4
pipreqs==0.4.11
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.20
proto-plus==1.19.5
protobuf==3.19.0
psutil==5.8.0
psycopg2-binary
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.11.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
pylint==2.6.0
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyrebase4==4.5.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==3.3.0
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.2.1
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.2
qtconsole==5.1.1
QtPy==1.10.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.19.0
rsa==4.7.2
Rtree==0.9.7
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0
scikit-surprise==1.1.1
scipy==1.7.1
setuptools==58.0.4
sip==5.0.0
six==1.15.0
snowballstemmer==2.1.0
sonopy==0.1.2
sortedcontainers==2.4.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.26
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==3.0.0
three-merge==0.1.1
tinycss==0.4
tokenizers==0.10.3
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.11.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
ujson==4.0.2
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.7
vader==0.0.2
vaderSentiment==3.3.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
whichcraft==0.6.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
zipp==3.6.0

